舵机加减速操作交由底层驱动进行

+++ task/update.sh
@@ -1,9 +1,16 @@
 #! /bin/sh
 
-rm /home/tao/work/linux-3.0-rc6-Old/drivers/char/taoTower.ko
-cd /home/tao/work/linux-3.0-rc6-Old/
-make SUBDIRS=drivers/char/ modules
+#使用当前目录下的底层驱动源码，放到内核源码下编译出模块．
+cmp ./taoTower.c /home/tao/work/linux-3.0-rc6-Old/drivers/char/taoTower.c
+if [ $? ]
+then
+    rm /home/tao/work/linux-3.0-rc6-Old/drivers/char/taoTower.ko
+    cp ./taoTower.c /home/tao/work/linux-3.0-rc6-Old/drivers/char/ 
+    cd /home/tao/work/linux-3.0-rc6-Old/
+    make SUBDIRS=drivers/char/ modules
+fi
 
+#应用层编译
 cd /home/tao/program_practice/graduate_designed/task/
 make
